+++ api/src/pdc/services/policy/engine/policies/index.ts
@@ -27,7 +27,7 @@ import { PmgfLate2023 } from './20231001_PMGF';
 import { SMT2022 } from './20220414_SMT';
 import { SMT2023 } from './20230126_SMT';
 import { SMTC2024 } from './20240101_SMTC';
-import { TerresTouloises2024 } from './TerresTouloises2024';
+import { TerresTouloises2024 } from './20240201_TerresTouloises';
 import { Vitre2023 } from './Vitre2023';
 import { PolicyTemplateOne } from './unbound/PolicyTemplateOne';
 import { PolicyTemplateThree } from './unbound/PolicyTemplateThree';
